Fix infinite loop on generate spell shop item
Crash was caused on conditions:
- We generate a shop item for a spell ->
- That spell has at least one modifier (random range is 1/2, not 0)

The unique modifier loop would then run forever, as the condition for it to continue was that the spell's modifier list did **NOT** contain the chosen modifier, which would always be true since the spell has no modifiers to begin with. A valid fix for the bug would've also just been to remove the `!` in the while condition, but I've rewritten it to use LINQ as I feel it conveys intent better.