Replace UI Rounded Corners package
UI Rounded Corners, the package used to generate rounded UI corners, has been replaced with a new package for rounding UI corners, called Rounded Corners (not to be confused with UI Rounded Corners), which rounds the corners of UI.

Jokes aside, it uses a different method: directly modifying the image's mesh as opposed to using generated materials, which should avoid so many merge conflicts with scenes. Should be more or less visually identical.